update repl package to work with latest sveltekit
